Fix bikeshed warnings
git-svn-id: http://google-web-toolkit.googlecode.com/svn/branches/2.1@8128 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ bikeshed/src/com/google/gwt/valuestore/shared/impl/RecordJsoImpl.java
@@ -83,7 +83,7 @@ public class RecordJsoImpl extends JavaScriptObject implements Record {
       return (V) Double.valueOf(getDouble(property.getName()));
     }
     if (Date.class.equals(property.getType())) {
-      double millis = (double) new Date().getTime();
+      double millis = new Date().getTime();
       if (isDefined(property.getName())) {
         millis = getDouble(property.getName());
       }
@@ -159,7 +159,7 @@ public class RecordJsoImpl extends JavaScriptObject implements Record {
       return;
     }
     if (value instanceof Date) {
-      double millis = (double) ((Date) value).getTime();
+      double millis = ((Date) value).getTime();
       setDouble(property.getName(), millis);
       return;
     }
